refactor: move series utils to types/xy_chart/utils
