+++ src/lib/series/series.ts
@@ -2,10 +2,10 @@ import { ColorConfig } from '../../utils/themes/theme';
 import { Accessor } from '../../utils/accessor';
 import { GroupId, SpecId } from '../../utils/ids';
 import { splitSpecsByGroupId, YBasicSeriesSpec } from './domains/y_domain';
-import { formatNonStackedDataSeriesValues } from './nonstacked_series_utils';
-import { isEqualSeriesKey } from './series_utils';
+import { formatNonStackedDataSeriesValues } from '../../types/xy_chart/utils/nonstacked_series_utils';
+import { isEqualSeriesKey } from '../../types/xy_chart/utils/series_utils';
 import { BasicSeriesSpec, Datum, SeriesAccessors } from './specs';
-import { formatStackedDataSeriesValues } from './stacked_series_utils';
+import { formatStackedDataSeriesValues } from '../../types/xy_chart/utils/stacked_series_utils';
 
 export interface RawDataSeriesDatum {
   /** the x value */
