+++ src/types/xy_chart/rendering/rendering.ts
@@ -15,7 +15,7 @@ import { Scale, ScaleType } from '../../../utils/scales/scales';
 import { CurveType, getCurveFactory } from '../../../utils/curves';
 import { LegendItem } from '../../../lib/series/legend';
 import { DataSeriesDatum } from '../../../lib/series/series';
-import { belongsToDataSeries } from '../../../lib/series/series_utils';
+import { belongsToDataSeries } from '../utils/series_utils';
 import { DisplayValueSpec } from '../../../lib/series/specs';
 
 export interface GeometryId {
